test(filesystem): fold end-to-end delete tests into fs integration test
Move the deterministic full-agent delete tests out of the standalone
fs.delete.e2e.test.ts and into a 'delete tool' describe within
fs.int.test.ts, alongside the other createDeepAgent integration tests. Same
four scenarios (single-file success, recursive directory delete, missing-path
error, permission-denied preserves state), driven by a FakeListChatModel delete
tool call. No new file.
